fix: adjusted related file names with route path
