+++ app/routes.ts
@@ -27,7 +27,6 @@ export default [
 	route("/speaker", "routes/speaker.tsx"),
 	route("/organizers", "routes/organizers.tsx"),
 	route("/organizer", "routes/organizer.tsx"),
-	route("/user-profile", "routes/user-profile.tsx"),
 	layout("routes/layouts/auth.tsx", [
 		route("/login", "routes/login.tsx"),
 		route("/register", "routes/register.tsx"),
@@ -38,8 +37,8 @@ export default [
 	...prefix("auth", [
 		layout("routes/layouts/protected.tsx", [
 			route("/logout", "routes/auth/logout.tsx"),
-			route("/user-profile", "routes/auth/dashboard.tsx"),
-			route("/dashboard", "routes/auth/user-profile.tsx"),
+			route("/user-profile", "routes/auth/user-profile.tsx"),
+			route("/dashboard", "routes/auth/dashboard.tsx"),
 			route("/payment", "routes/auth/payment.tsx"),
 			route("/user-ticket", "routes/auth/user-ticket.tsx"),
 			route("/check-in", "routes/auth/check-in.tsx"),
